requirements.txt
fastapi
uvicorn
joblib
numpy
pillow
opencv-python-headless
jinja2
python-multipart
